apps/jpkr-admin/api/pyproject.toml
[tool.poetry]
name = "onigiri-jpkr-admin"
version = "0.0.1"
description = "Onigiri Japanese-Korean App Admin API"
authors = ["Jaehak Lee <leejaehak87@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
sqlalchemy = "^2.0.41"
uvicorn = "^0.34.3"
fastapi = "^0.115.14"
pgvector = "^0.4.1"
dotenv = "^0.9.9"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.1"
requests = ">=2.32"
fugashi = {extras = ["unidic-lite"], version = "^1.5.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
google-auth = ">=2.30"
google-auth-oauthlib = ">=1.2"
psycopg2-binary = "^2.9.10"
boto3 = "^1.40.19"
pydantic-settings = "^2.10.1"
torch = {version = "==2.6.0", source = "pytorch_gpu"}
torchvision = {source = "pytorch_gpu"}
torchaudio = {source = "pytorch_gpu"}
sentence-transformers = "==2.7.0"
sentencepiece = "^0.2.0"
pyopenjtalk = "^0.4.1"
espnet = {path = "espnet", develop = true}
flash-attn = {path = "flash_attn-2.8.3+cu124torch2.6.0cxx11abiFALSE-cp311-cp311-win_amd64.whl"}
espnet-model-zoo = "^0.1.7"
pydub = "^0.25.1"
transformers = "==4.44.2"
accelerate = "^1.10.1"

[[tool.poetry.source]]
name = "pytorch_gpu"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"